# packages/openapi.json
{
  "components": {
    "schemas": {
...
      "CreatePATRequest": {
        "properties": {
          "name": {
            "type": "string"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "user_id": {
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "name"
        ],
        "type": "object"
      },
      "CreatePATResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "token": {
            "type": "string"
          },
          "user_id": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "user_id",
          "token",
          "scopes",
          "created_at"
        ],
        "type": "object"
      },
...
      "ImportRequest": {
        "properties": {
          "entities": {
            "items": {
              "properties": {
                "display_name": {
                  "type": "string"
                },
                "identifier": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                },
                "profile": {
                  "additionalProperties": {},
                  "type": "object"
                },
                "schema_id": {
                  "type": "string"
                },
                "state": {
                  "type": "string"
                }
              },
              "type": "object"
            },
            "type": "array"
          },
          "linked_identities": {
            "items": {
              "properties": {
                "external_email": {
                  "type": "string"
                },
                "external_sub": {
                  "type": "string"
                },
                "provider_name": {
                  "type": "string"
                },
                "user_identifier": {
                  "type": "string"
                }
              },
              "type": "object"
            },
            "type": "array"
          },
          "on_conflict": {
            "type": "string"
          },
          "providers": {
            "items": {
              "properties": {
                "auto_register": {
                  "type": "boolean"
                },
                "claim_overrides": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                },
                "config": {
                  "additionalProperties": {},
                  "type": "object"
                },
                "id": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "protocol": {
                  "type": "string"
                },
                "template": {
                  "type": "string"
                }
              },
              "type": "object"
            },
            "type": "array"
          }
        },
        "required": [
          "on_conflict"
        ],
        "type": "object"
      },
      "ImportResult": {
        "properties": {
          "id": {},
          "index": {
            "type": "integer"
          },
          "reason": {
            "type": "string"
          },
          "resource": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "index",
          "resource",
          "status"
        ],
        "type": "object"
      },
      "ListResponse": {
        "properties": {
          "items": {},
          "next_cursor": {
            "type": "string"
          },
          "total": {
            "type": "integer"
          }
        },
        "required": [
          "items"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "PAT or session token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Identity and access management API. Schema-driven, extensible, open source.",
    "title": "Zitadel API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/admin/bulk": {
      "post": {
        "operationId": "adminBulk",
        "responses": {
          "200": {
            "description": "Bulk create entities (admin)"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Bulk create entities (admin)",
        "tags": [
          "Admin"
        ]
      }
    },
...
    "/v1/import": {
      "post": {
        "operationId": "importData",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResult"
                }
              }
            },
            "description": "Import entities and providers"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Import entities and providers",
        "tags": [
          "Import"
        ]
      }
    },
...
    "/v1/pats": {
      "get": {
        "operationId": "listPATs",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse"
                }
              }
            },
            "description": "List personal access tokens"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List personal access tokens",
        "tags": [
          "PATs"
        ]
      },
      "post": {
        "operationId": "createPAT",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePATRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatePATResponse"
                }
              }
            },
            "description": "Create a personal access token"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create a personal access token",
        "tags": [
          "PATs"
        ]
      }
    },
    "/v1/pats/{id}": {
      "delete": {
        "operationId": "revokePAT",
        "parameters": [
          {
            "description": "Resource ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Revoke a personal access token"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Revoke a personal access token",
        "tags": [
          "PATs"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "This instance",
      "url": "/"
    }
  ],
...
}